Fix flickering black border
Summary:
This addresses an issue in transitioning from the vertical to the box cursor whereby the CSS
transition gets inappropriately applied, causing three edges of the box to start off black before
transitioning to green. Instead, we should set all border sides to default to the active color, so
that in the event of a transition from the vertical to the box cursor, we don't animate the new
edges from black to the active color.

Test Plan:
- Type in any math that triggers a box cursor (exponent, fraction).
- Type some math to make the box cursor disappear.
- Delete until the box cursor returns.
    - Verify that it starts off green on all edges and is never black.
    - (Previously, the cursor would start off black on the top, right, and bottom edges before animating to green).

Auditors: shadaj