Adjust button sizing on configuration change
Summary:
This fixes a bug whereby changing the keypad type wouldn't force a layout pass for its keys. This
was never an issue in the playground, since you can't switch keypad types and the 'initial' type is
always the correct type, but when testing in webapp, I noticed that exercises that used fraction
keypads would be laid out in too narrowly.

Test Plan:
- Substate into Perseus; substate into webapp.
- Load an exercise with a numeric input (e.g., 'Decimals in words').
- Tap on the input.
    - Verify that the keypad is rendered at full-width.

Reviewers: kevinb, jared

Reviewed By: jared

Differential Revision: https://phabricator.khanacademy.org/D29431