Account for persistent bottom bar when scrolling input into view
Summary:
Now that the bottom is persistent and below the keypad, our logic for scrolling an input into view
is faulty and, in fact, will always leave the input obscured by the keypad. This diff incorporates
the bottom bar's presence into the scrolling logic. It's a hacky fix (which won't be precise in the
native apps), but feels like a quick and big win for useability.

On Asana @ https://app.asana.com/0/132409062806753/162356734461122.

Test Plan:
- Substate into webapp.
- Open an exercise.
- Scroll such that an input would be below the keypad.
- Tap on the input.
    - Verify that it's scrolled into view properly.

{F549854}

Reviewers: benkomalo, kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D29570