Avoid dismissing keypad on horizontal scrolls
Summary:
Previously, swiping in the content area in any way that didn't cause a page scroll would lead to
the keypad's dismissal. Now, we track any touches and avoid dismissing the keypad if any of the
user's fingers moved more than a specified distance (8px, which is the default touch slop in the
Android SDK).

On Asana @ https://app.asana.com/0/132409062806753/162277152320498.

Test Plan:
- Open up the playground.
- Tap to open the keypad.
- Tap elsewhere to dismiss it.
- Re-open the keypad.
- Tap and drag your finger slightly.
    - Verify that the keypad does not dismiss.

Reviewers: kevinb

Reviewed By: kevinb

Subscribers: benkomalo

Differential Revision: https://phabricator.khanacademy.org/D29677