Apply fill-rule fix when icon updates
Summary:
We have this fill-rule fix that needs to be applied to our SVG icons. Now that we have dynamic
icons, we need to make sure that we run the fix whenever the component updates with a new icon.
The bug here (that the icons appear filled rather than hollow) didn't manifest itself locally
because we run React 15 in the playground, and this fix is for React 14 (which webapp runs).

Test Plan:
- Substate into Perseus; substate into webapp.
- Open an exercise on phones.
- Open the keypad.
- Create a fraction.
    - Verify that the icon is hollow and not filled.

Auditors: kevinb