Stop propagation for click/touchend
Summary:
This makes it so that when you tap an input that's in a zoomable, the zoomable
doesn't also zoom back out.

Asana: https://app.asana.com/0/47726783949319/144457874535340
Will require a small change to perseus to work 100%

Test Plan:
Fire up:
http://localhost:8081/math/algebra/systems-of-linear-equations/solving-systems-of-equations-with-substitution/e/systems_of_equations_with_substitution
Tap the table, it should zoom in
Tap the input, it should select, and the table should *not* zoom back out.

This table:
{F550614}

GIF
{F550739}

Reviewers: charlie, kevinb

Reviewed By: kevinb

Subscribers: benkomalo

Differential Revision: https://phabricator.khanacademy.org/D29888