Fix held down backspace on Android devices
Summary:
https://app.asana.com/0/143216769809283/168196116149999

This fixes an issue that was reproducing fairly reliably for me:
pressing backspace in a math keypad would always have it be held down.

It turns out we were getting multiple touch starts from the webview
and we weren't robust to this. This fixes that.

Test Plan:
 - point a local android device to load from localhost
 - open any exercise (e.g. multiply by 10s)
 - type in some numbers
 - press back, see it erase one and only one
 - press and hold, see that still work
 - type in other numbers, see they still get inputted

also run the unit test

Reviewers: jared, kevinb

Reviewed By: kevinb

Subscribers: charlie

Differential Revision: https://phabricator.khanacademy.org/D30100