Hardware keyboard input (#36)
Adds the ability to use a hardware keyboard with the math input. This allows basic input of numbers, operators, and the delete key.

In the future we could choose to allow multi character input (like != or >=) and complex input (like sin, cos, tan), and more movement keys (like arrows, control+a, control+e). All of these things will greatly increase the complexity in the implementation and might be platform dependent so we want to wait until learners give a clear signal that they want these features.