Show overflow input values when typing or pulling cursor handle
SUMMARY
Issue: https://khanacademy.atlassian.net/browse/LP-7812

It turns out that if we remove the code that overrides Mathquil's native scrolling logic, the extra scrolly behavior we want (scrolling on cursor handle pull and when the user is typing past the edge of the box) just magically work. For some reason, we still needed to add the `overflow: "scroll"` styles in `override.less` to get manually scrolling to work, but Mathquil can handle the other stuff for us.

So this diff mostly just removes the code that overrides Mathquil's code. That also means removing the `scrollable` prop, which I really should have removed before anyway, since we're now making the input field always scrollable. (Should we be concerned about removing this prop since this is an open source project? I actually don't have any experience with open source, so I really don't know.)

I also updated my code for hiding the cursor handle on scroll since that currently messes up the feature of scrolling by pulling on the cursor handle (for obvious reasons). This isn't actually part of the ticket I set out to do, but since Mathquil is giving it to us for free, why not? It's not perfect though. It works pretty well when scrolling the input field left by pulling on the cursor, but scrolling it right is really janky. So I was thinking of taking the ticket I had for implementing the scroll-on-handle-pull feature and, instead of closing it, updating it to instead say to make this feature better. But since it's already a P3, I don't know if we'll ever get to it, so maybe it's better to just close it? https://khanacademy.atlassian.net/browse/LP-7695

TEST PLAN
Open the expression widget,
type in a whole buncha values,
when you pass the max width of the input field, verify the input field scrolls to show the new values you insert,
verify you can manually scroll left and right by using a track pad, scroll wheel, or by pulling the scroll bar,
tap into the input field,
verify the cursor handle appears,
scroll the field again,
verify the handle disappears,
tap into the field again,
grab the cursor handle,
verify you can pull it around the input field,
pull it off the side of the input field in a direction that has hidden values,
and verify the input field scrolls to show the hidden values.