Make input field aria label more specific (#78)
* Make input field aria label more specific

We have an issue where on Android mobile devices, you need to use two fingers to tap on the input field to make the keyboard appear. It should only require one finger to tap, which it does on iOS. We don't have time right now to figure out why this isn't working, so, for the meantime, we want to update the aria label to communicate that if one finger tapping isn't working, try two fingers.

There is another issue, not addressed here, where the aria label doesn't seem to work when you tap on the input field. It only activates if you navigate to the input field via swipes. So sighted users who are on an android device and using talkback won't hear this instruction. I don't know when we're going to deal with that. I might have to do that before closing the associated ticket.

I also made some changes to how Tabbar is being imported because it was breaking the tests.