FEI-4197: Update react-redux to 7.2.6 (#95)
## Summary:
FEI-4163 is updating webapp to use react-redux 7.2.6 so that we can use the hooks that are available in that future.  In order to ensure that math-input will continue to work after this upgrade we also need to update react-redux here as well.  One of the changes that was made in react-redux was for connect() to return a functional component.  There are a few places where we use refs with connect() wrapped components.  Thankfully, in the new version of react-redux, connect() has an option for it to use forwardRef() so that we can use refs with the components it returns.  I've updated all calls of connect() to use this option so we don't have to worry about tracking down which places it's actually needed.

I also ended up having to upgrade Flow due to it reporting an error in the flow types from some of the wonder-blocks components.

Issue: FEI-4197

## Test plan:
**local testing**
- yarn start
- load http://localhost:9000/ in mobile/responsive mode in your web browser
- tap on the input field, see the keypad show up at the bottom
- type some keys, see the numbers/symbols appear in the input
- click on the input again, drag the cursor handle
- tap the dismiss button on the keypad, see the keypad disappear

**webapp testing**
- yarn build
- yarn pack
- cp khanacademy-math-input.tgz to webapp/services/static/node_modules/@khanacademy
- decompress the tarball and rename it to `math-input`
- run the item-renderer_test.js from webapp/service/static/

Author: kevinbarabash

Reviewers: kevinbarabash, jeresig, nixterrimus, somewhatabstract

Required Reviewers:

Approved By: jeresig

Checks: ✅ Mixed content - khan-math-input, ⚪️ Redirect rules - khan-math-input, ⚪️ Header rules - khan-math-input, ⚪️ Pages changed - khan-math-input, ✅ typecheck, ✅ test

Pull Request URL: https://github.com/Khan/math-input/pull/95